# Conflicts:
#	app/src/main/java/com/example/ding/umutos/presentation/LoginActivity.java

+++ app/src/main/java/com/example/ding/umutos/presentation/LoginActivity.java
@@ -102,7 +102,8 @@ public class LoginActivity extends AppCompatActivity {
             int count;
 
             File outFile = new File(copyPath);
-            //if (!outFile.exists()){
+
+            if (!outFile.exists()){
 
                 InputStreamReader in = new InputStreamReader(assetManager.open(asset));
                 FileWriter out = new FileWriter(outFile);
@@ -115,7 +116,7 @@ public class LoginActivity extends AppCompatActivity {
 
                 out.close();
                 in.close();
-            //}
+            }
 
             BufferedReader br = new BufferedReader(new FileReader(outFile));
